+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Group changes to describe their impact on the project, as follows:
 ## [6.1.0] - Unreleased
 
 ### Added
+- Added the ability to edit/delete chat messages sent less than 24 hours ago.
 - Added hover effect when using a mouse (useful for tablets or devices with desktop mode)
 - Support right click on some items to open bottom sheet/menu
 - Added toggle speaker action in active call notification